Done changes related to validate bulk uploaded questions
