+++ src/app/bulk/components/bulk-upload/bulk-upload.component.scss
@@ -1,3 +1,7 @@
 ul {
   list-style: square inside
 }
+
+.error { 
+  color: red;
+}
