+++ src/app/bulk/components/bulk-upload/bulk-upload.component.ts
@@ -22,7 +22,8 @@ export class BulkUploadComponent implements OnInit, OnDestroy {
 
   tagsObs: Observable<string[]>;
   categoriesObs: Observable<Category[]>;
-  parseError: boolean;
+  parseError = false;
+  parseErrorMessage: String;
   bulkUploadFileInfo: BulkUploadFileInfo;
   file: File;
 
@@ -93,33 +94,62 @@ export class BulkUploadComponent implements OnInit, OnDestroy {
   generateQuestions(csvString: string): void {
     parse(csvString, { 'columns': true, 'skip_empty_lines': true },
       (err, output) => {
-        this.questions =
-          output.map(element => {
-            const question: Question = new Question();
-            question.questionText = element['Question'];
-            question.answers = [
-              { 'id': 1, 'answerText': element['Option 1'], correct: false },
-              { 'id': 2, 'answerText': element['Option 2'], correct: false },
-              { 'id': 3, 'answerText': element['Option 3'], correct: false },
-              { 'id': 4, 'answerText': element['Option 4'], correct: false }
-            ]
-            question.answers[element['Answer Index'] - 1].correct = true;
-            question.tags = [];
+        if (output !== undefined && output !== '') {
+          this.questions =
+            output.map(element => {
 
-            for (let i = 1; i < 10; i++) {
-              if (element['Tag ' + i] && element['Tag ' + i] !== '') {
-                question.tags.push(element['Tag ' + i]);
+              console.log(element);
+
+              const question: Question = new Question();
+              question.questionText = element['Question'];
+              question.answers = [
+                { 'id': 1, 'answerText': element['Option 1'], correct: false },
+                { 'id': 2, 'answerText': element['Option 2'], correct: false },
+                { 'id': 3, 'answerText': element['Option 3'], correct: false },
+                { 'id': 4, 'answerText': element['Option 4'], correct: false }
+              ]
+
+              if (question.answers[element['Answer Index'] - 1] !== undefined) {
+                question.answers[element['Answer Index'] - 1].correct = true;
               }
-            }
+              question.tags = [];
 
-            question.published = false;
-            question.explanation = 'status - not approved';
-            question.status = QuestionStatus.PENDING;
-            question.created_uid = this.user.userId;
-            return question;
-          });
+              for (let i = 1; i < 10; i++) {
+                if (element['Tag ' + i] && element['Tag ' + i] !== '') {
+                  question.tags.push(element['Tag ' + i]);
+                }
+              }
 
-        this.bulkUploadFileInfo.uploaded = this.questions.length;
+              question.published = false;
+              question.explanation = 'status - not approved';
+              question.status = QuestionStatus.PENDING;
+              question.created_uid = this.user.userId;
+
+              if (!question.questionText || question.questionText.trim() === '') {
+                this.parseError = true;
+                this.parseErrorMessage = 'Missing Question';
+              } else if (question.answers[0].answerText.trim() === '' || question.answers[1].answerText.trim() === '' ||
+                question.answers[2].answerText.trim() === '' || question.answers[3].answerText.trim() === '') {
+                this.parseError = true;
+                this.parseErrorMessage = 'Missing Question Answer Options';
+              } else if (question.answers.filter(a => a.correct).length !== 1) {
+                this.parseError = true;
+                this.parseErrorMessage = 'Must have exactly one correct answer';
+              } else if (question.answers.filter(a => !a.answerText || a.answerText.trim() === '').length > 0) {
+                this.parseError = true;
+                this.parseErrorMessage = 'Missing Answer';
+              } else if (question.tags.length < 3) {
+                this.parseError = true;
+                this.parseErrorMessage = 'Not enough tags';
+              }
+
+              return question;
+            });
+          this.bulkUploadFileInfo.uploaded = this.questions.length;
+        } else {
+          this.parseError = true;
+          this.parseErrorMessage = 'File format is not correct';
+        }
       });
   }
