+++ src/app/shared/components/question-form/question-form.component.scss
@@ -1,0 +1,77 @@
+
+#curentactive { text-align: center; }
+.submitpage .card{ padding: 25px 0 0 }
+.optionbox{ margin-top: 30px; text-align: left;}
+.optionbox h2{ font-size: 20px; letter-spacing: 0.75px; line-height: 1.2; margin: 0 0 10px }
+.optionbox h2 span{ font-size: 20px; text-transform: none; letter-spacing: .10px; padding-left: 0; font-family: 'hindmedium' }
+.submitpage .optionbox{ padding: 0 10px }
+.submitpage textarea{ width: 100%; display: inline-block; border-width: 1px; border-style: solid; border-radius: 4px; font-size: 18px; }
+.card .enterquestion h3{ font-size: 18px; margin: 30px 0 16px}
+.enterquestion .optionbox{ margin-top: 0 }
+.card .enterquestion h3:before{ display: none;  }
+
+.enterquestion .formfield{ padding: 0; margin-bottom: 32px; }
+.enterquestion .formfield > input{ border-width: 1px; border-style: solid; border-radius: 4px; padding: 10px 25px; font-size: 18px; display: inline-block; width: calc(100% - 34px)}
+.formfield select {  border-width: 1px; border-style: solid; border-radius: 4px; font-size: 18px; display: inline-block; }
+.enterquestion .formfield > label{ width: 30px; display: inline-block; vertical-align: -4px}
+.enterquestion .formfield input[type="radio"]{ display: block;}
+
+.question-list li input[type="radio"], .optionbox input[type="radio"]{ display: none; }
+.question-list li label{ cursor: pointer; font-family: 'hindmedium'; font-size: 16px; padding: 6px 13px; border-radius: 17px; transition: 0.3s ease-in-out; position: relative; display: inline-block;}
+.question-list li label:before{ content: "";position: absolute; height: 17px; width: 8px; border-width:0 2px 2px 0; border-style: solid; transform: rotate(45deg); left: 22px; display: none;}
+.question-list li input:checked ~ label{ padding-left: 40px; }
+.question-list li input:checked ~ label:before{ display: block; }
+#curentactive .btnbox{ background: #fff; }
+.btn.startg{ font-size: 24px; display: block; padding: 15px 0; width: 100% ; max-width: 300px;line-height: 1.2; margin:0 auto}
+.btn.startg:disabled{ background-color: silver; }
+@media screen and (min-width: 767px){
+  .optionbox h2 span{ font-size: 22px }
+  .pcategories .formfield select{  width: 300px }
+  .card .enterquestion h3{margin-top: 40px}
+  .formfield, .mysetting .formfield { display: inline-block; }
+  .savebox a.btn.startg{ display: inline-block; margin-right: 70px}
+  .enterquestion .formfield{ width: calc(50% - 3px) }
+  .enterquestion .formfield:nth-child(2n + 1){ text-align: right; }
+  .enterquestion .formfield > input{width: calc(85%) }
+}
+/*
+//commenting this out as @ 1024px the menu appears; unless we want to add max-width of 1024 here.
+@media screen and (min-width: 1000px){
+  .optionbox{width: 765px; margin: 60px auto 0;}  
+  .optionbox h2{ font-size: 16px; text-transform: uppercase; margin-bottom: 0 }
+  .optionbox .pcategories h2{ margin-top: 40px}
+  .submitpage .optionbox{ margin-top: 0; padding: 0 }
+  .submitpage .optionbox h2{ margin-bottom: 7px }  
+  .btn.startg{ width: 216px; margin:0 auto; padding: 20px 0  }
+  .submitpage .enterquestion .optionbox h2{ margin-bottom: 15px }
+}
+*/
+@media screen and ( max-width:760px){
+  .savebox a.btn.startg{ width: 90%; margin: 0 auto 30px; }
+}
+
+
+mat-card {
+  margin: 15px;
+}
+mat-card-title {
+  margin-bottom: 50px;
+}
+mat-card-actions {
+  margin-bottom: 0px;
+}
+//.question-form {
+//  width: 90%;
+//}
+.full-width {
+  width: 100%;
+}
+.remove-tag {
+  cursor: pointer;
+}
+::ng-deep mat-input-container.ng-invalid.ng-touched .mat-input-underline {
+  border-color: red;
+}
+::ng-deep mat-hint.error, .error {
+  color: red;
+}
