+++ src/app/shared/shared.module.ts
@@ -10,7 +10,7 @@ import { RouterModule } from '@angular/router';
 
 import {
   QuestionComponent, QuestionsComponent, QuestionsSearchComponent, QuestionsTableComponent,
-  BulkSummaryTableComponent, BulkSummaryQuestionComponent
+  BulkSummaryTableComponent, BulkSummaryQuestionComponent, QuestionFormComponent
 } from './components';
 
 @NgModule({
@@ -20,7 +20,8 @@ import {
     QuestionsSearchComponent,
     QuestionsTableComponent,
     BulkSummaryTableComponent,
-    BulkSummaryQuestionComponent
+    BulkSummaryQuestionComponent,
+    QuestionFormComponent
   ],
   imports: [
     CommonModule,
@@ -47,7 +48,7 @@ import {
   ],
   exports: [QuestionComponent, QuestionsComponent, QuestionsSearchComponent, QuestionsTableComponent,
     BulkSummaryTableComponent, BulkSummaryQuestionComponent, CommonModule, HttpClientModule, ReactiveFormsModule,
-    FlexLayoutModule,
+    FlexLayoutModule, QuestionFormComponent,
     SharedMaterialModule, CdkTableModule]
 })
 export class SharedModule { }
